PR review refactor1.
